добавил alembic.

+++ order_api_service/source/database/models/product_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, Boolean
+from sqlalchemy import Column, String, Integer, Boolean, Float
 from sqlalchemy.orm import relationship
 
 from .base_model import BaseModel
@@ -17,7 +17,7 @@ class Product(BaseModel):
     product_stripe_id = Column(String(32))  # id продукта в Stripe
     description = Column(String(255))
     duration = Column(Integer, nullable=False)  # в месяцах
-    price = Column(Integer, nullable=False)
+    price = Column(Float, nullable=False)
     currency_code = Column(String(3), default='rub')
     recurring = Column(Boolean, default=True)
     is_active = Column(Boolean, default=True)
